+++ src/main/java/com/easypost/model/Field.java
@@ -4,7 +4,6 @@ import lombok.Getter;
 
 @Getter
 public class Field extends EasyPostResource {
-    private String key;
     private String visibility;
     private String label;
     private String value;