+++ src/test/java/com/easypost/CarrierAccountTest.java
@@ -2,12 +2,17 @@ package com.easypost;
 
 import com.easypost.exception.API.InvalidRequestError;
 import com.easypost.exception.EasyPostException;
+import com.easypost.http.Requestor;
 import com.easypost.model.CarrierAccount;
 import com.easypost.model.CarrierType;
+import com.easypost.model.Pickup;
+import com.easypost.model.Shipment;
 import com.google.common.collect.ImmutableMap;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
+import org.mockito.MockedStatic;
+import org.mockito.Mockito;
 
 import java.util.HashMap;
 import java.util.List;
@@ -22,6 +27,8 @@ public final class CarrierAccountTest {
     private static String testCarrierAccountId = null;
 
     private static TestUtils.VCR vcr;
+
+    private static MockedStatic<Requestor> requestMock = Mockito.mockStatic(Requestor.class);
 
     /**
      * Set up the testing environment for this file.
@@ -178,5 +185,59 @@ public final class CarrierAccountTest {
 
         assertInstanceOf(List.class, types);
         assertTrue(types.stream().allMatch(type -> type != null));
+    }
+
+    /**
+     * Test that the CarrierAccount fields are correctly deserialized from the API response.
+     * None of the demo carrier accounts used in the above tests have credentials or test credentials fields,
+     * so we need to use some mock data.
+     */
+    @Test
+    public void testCarrierFieldsJsonDeserialization() {
+        String carrierAccountJson = "[{\"id\":\"ca_123\",\"object\":\"CarrierAccount\"," +
+                "\"fields\":{\"credentials\":{\"account_number\":{\"visibility\":\"visible\"," +
+                "\"label\":\"DHL Account Number\",\"value\":\"123456\"},\"country\":{\"visibility\":\"visible\"," +
+                "\"label\":\"Account Country Code (2 Letter)\",\"value\":\"US\"},\"site_id\":{\"visibility\":" +
+                "\"visible\",\"label\":\"Site ID (Optional)\",\"value\": null },\"password\":{\"visibility\":" +
+                "\"password\",\"label\":\"Password (Optional)\",\"value\":\"\"},\"is_reseller\":{\"visibility\":" +
+                "\"checkbox\",\"label\":\"Reseller Account? (check if yes)\",\"value\":null}}}}]";
+        CarrierAccount[] carrierAccounts = Constants.Http.GSON.fromJson(carrierAccountJson, CarrierAccount[].class);
+
+        CarrierAccount carrierAccount = carrierAccounts[0];
+        assertEquals("ca_123", carrierAccount.getId());
+        assertEquals("CarrierAccount", carrierAccount.getObject());
+        assertEquals("DHL Account Number",
+                carrierAccount.getFields().getCredentials().get("account_number").getLabel());
+    }
+
+    /**
+     * Test that the CarrierAccount fields are correctly serialized to the API request.
+     */
+    @Test
+    public void testCarrierFieldsJsonSerialization() {
+        String carrierAccountJson = "[{\"id\":\"ca_123\",\"object\":\"CarrierAccount\",\"fields\":{\"credentials\":" +
+                "{\"account_number\":{\"visibility\":\"visible\",\"label\":\"DHL Account Number\"," +
+                "\"value\":\"123456\"},\"country\":{\"visibility\":\"visible\",\"label\":" +
+                "\"Account Country Code (2 Letter)\",\"value\":\"US\"},\"site_id\":{\"visibility\":\"visible\"," +
+                "\"label\":\"Site ID (Optional)\",\"value\": null },\"password\":{\"visibility\":\"password\"," +
+                "\"label\":\"Password (Optional)\",\"value\":\"\"},\"is_reseller\":{\"visibility\":\"checkbox\"," +
+                "\"label\":\"Reseller Account? (check if yes)\",\"value\":null}}}}]";
+        CarrierAccount[] carrierAccounts = Constants.Http.GSON.fromJson(carrierAccountJson, CarrierAccount[].class);
+        CarrierAccount carrierAccount = carrierAccounts[0];
+
+        // Prepare a parameter set for creating a pickup, using the carrier account object
+        Map<String, Object> pickupData = Fixtures.basicPickup();
+        pickupData.put("shipment", new Shipment());
+        pickupData.put("carrier_accounts", new CarrierAccount[] { carrierAccount });
+
+        // Avoid making a real request to the API, interested in pre-request serialization, not interested in response
+        requestMock.when(() -> Requestor.request(Requestor.RequestMethod.POST, "pickups", pickupData, Shipment.class,
+                vcr.client)).thenReturn(new Pickup());
+
+        // This will throw an exception if the carrier account fields could not be serialized properly
+        assertDoesNotThrow(() -> vcr.client.pickup.create(pickupData));
+
+        // Close mock
+        requestMock.close();
     }
 }